+++ src/CHANGELOG.md
@@ -2,7 +2,16 @@
 
 Use inverse order
 
-2025-07-06 Andrew H. Fagg
+##2025-07-22 Andrew H. Fagg
+SuperDataSet:
+* All numpy folds and data sets are now represented as 4-tuples.  Nones are used if a tuple entry has no data.
+
+##2025-07-17 Andrew H. Fagg
+
+SuperDataSet:
+* Added args.data_columns_categorical_to_int_direct
+
+##2025-07-06 Andrew H. Fagg
 
 SuperDataSet:
 * Data files are loaded into self.data (refer to these as *data tables*)
@@ -14,3 +23,4 @@ SuperDataSet:
 Command Line Arguments:
 * Many arguments names have been changed to reflect the module that they refer to
 * Use of expired arguments now generates errors
+